Move project/Creates/palette.js to project/common
