+++ frontend/src/components/project/common/palettes.js
