+++ frontend/src/pages/ProjectPage.jsx
@@ -11,8 +11,8 @@ import ModalWindow from "@components/common/ModalWindow"
 import PageTitle from "@components/common/PageTitle"
 import Drawer from "@components/drawers/Drawer"
 import { ErrorBox } from "@components/errors/ErrorProjectPage"
-import { getProjectColor } from "@components/project/Creates/palettes"
 import Progress from "@components/project/common/Progress"
+import { getProjectColor } from "@components/project/common/palettes"
 import DrawerEdit from "@components/project/edit/DrawerEdit"
 import ProjectEdit from "@components/project/edit/ProjectEdit"
 import { SkeletonProjectPage } from "@components/project/skeletons/SkeletonProjectPage"
