+++ frontend/src/pages/settings/Account.jsx
@@ -6,7 +6,7 @@ import styled, { useTheme } from "styled-components"
 import Button, { ButtonGroup, buttonForms } from "@components/common/Button"
 import { LoaderCircleFull } from "@components/common/LoaderCircle"
 import ModalWindow from "@components/common/ModalWindow"
-import { getProjectColor } from "@components/project/Creates/palettes"
+import { getProjectColor } from "@components/project/common/palettes"
 import Color from "@components/project/edit/Color"
 import Error from "@components/settings/Error"
 import PasswordSection from "@components/settings/PasswordSection"
